# client/assets/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AHEM - Ad Hoc Email",
    "description": "Welcome to the AHEM API documentation.<br>AHEM provides an easy to use RESTful API that allows testing or building anything on top of the AHEM mail server.<br>All endpoints follow the URL: <br>https://www.ahem.email/api/<br><br>__Authorization__<br>A token must be obtained by calling https://www.ahem.email/auth/authenticate.<br>The response will contain a token field. Tokens must be added to each api call in the Authorization header and a Bearer token.",
    "contact": {},
    "termsOfService": "http://ahem-regtool.tk",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://ahem-regtool.tk",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/auth/token": {
      "post": {
        "tags": [
          "General"
        ],
        "summary": "Get access token",
        "description": "",
        "operationId": "GetAccessToken",
        "parameters": [
          {
            "name": "Content-Type",
            "in": "header",
            "description": "",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "application/json"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "text/plain": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true,
                    "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpcCI6Ijo6MSIsImlhdCI6MTUzOTU1NTI5NCwiZXhwIjoxNTM5NTU4ODk0fQ.Y8FhYY5OGbHNn6EOcrF81swX3xeLF4t7Q6F5L7z6Xdg"
                  },
                  "$ref": "#/components/schemas/GetAccessTokenResponse"
                },
                "example": {
                  "success": true,
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpcCI6Ijo6MSIsImlhdCI6MTUzOTU1NTI5NCwiZXhwIjoxNTM5NTU4ODk0fQ.Y8FhYY5OGbHNn6EOcrF81swX3xeLF4t7Q6F5L7z6Xdg"
                }
              }
            }
          }
        },
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
    "/api/mailbox/{mailboxName}/email": {
      "get": {
        "tags": [
          "Mailbox"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "get a mailbox's emails list",
        "description": "",
        "operationId": "GetAMailbox'sEmailsList",
        "parameters": [
          {
            "name": "mailboxName",
            "in": "path",
            "description": "mailbox name.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "example": "mailboxName"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "emailId": "5bc3c17cb76d8f4e1099725c",
                        "sender": {
                          "address": "test@my.domain.com",
                          "name": "AHEM Test!"
                        },
                        "subject": "AHEM mail test!",
                        "timestamp": 1539555708477,
                        "isRead": true
                      }
                    ],
                    "$ref": "#/components/schemas/GetAMailboxsEmailsListResponse"
                  },
                  "description": ""
                },
                "example": [
                  {
                    "emailId": "5bc3c17cb76d8f4e1099725c",
                    "sender": {
                      "address": "test@my.domain.com",
                      "name": "AHEM Test!"
                    },
                    "subject": "AHEM mail test!",
                    "timestamp": 1539555708477,
                    "isRead": true
                  }
                ]
              }
            }
          },
          "401": {
            "description": "Unexpected error in API call. See HTTP response body for details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAMailboxsEmailsList401Response"
                },
                "example": {
                  "success": false,
                  "message": "No token provided."
                }
              }
            }
          },
          "404": {
            "description": "Unexpected error in API call. See HTTP response body for details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAMailboxsEmailsList404Response"
                },
                "example": {
                  "error": "MAILBOX IS EMPTY!"
                }
              }
            }
          }
        },
        "x-operation-settings": {
          "CollectParameters": false,
          "AllowDynamicQueryParameters": false,
          "AllowDynamicFormParameters": false,
          "IsMultiContentStreaming": false
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "GetAccessTokenResponse": {
        "title": "Get access token response",
        "required": [
          "success",
          "token"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpcCI6Ijo6MSIsImlhdCI6MTUzOTU1NTI5NCwiZXhwIjoxNTM5NTU4ODk0fQ.Y8FhYY5OGbHNn6EOcrF81swX3xeLF4t7Q6F5L7z6Xdg"
          }
        },
        "example": {
          "success": true,
          "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpcCI6Ijo6MSIsImlhdCI6MTUzOTU1NTI5NCwiZXhwIjoxNTM5NTU4ODk0fQ.Y8FhYY5OGbHNn6EOcrF81swX3xeLF4t7Q6F5L7z6Xdg"
        }
      },
      "GetAMailboxsEmailsListResponse": {
        "title": "get a mailbox's emails list response",
        "required": [
          "emailId",
          "sender",
          "subject",
          "timestamp",
          "isRead"
        ],
        "type": "object",
        "properties": {
          "emailId": {
            "type": "string",
            "example": "5bc3c17cb76d8f4e1099725c"
          },
          "sender": {
            "type": "object",
            "example": {
              "address": "test@my.domain.com",
              "name": "AHEM Test!"
            }
          },
          "subject": {
            "type": "string",
            "example": "AHEM mail test!"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "example": 1539555708477
          },
          "isRead": {
            "type": "boolean",
            "example": true
          }
        },
        "example": {
          "emailId": "5bc3c17cb76d8f4e1099725c",
          "sender": {
            "address": "test@my.domain.com",
            "name": "AHEM Test!"
          },
          "subject": "AHEM mail test!",
          "timestamp": 1539555708477,
          "isRead": true
        }
      },
      "GetAMailboxsEmailsList404Response": {
        "title": "get a mailbox's emails list 404 response",
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "MAILBOX IS EMPTY!"
          }
        },
        "example": {
          "error": "MAILBOX IS EMPTY!"
        }
      },
      "GetAMailboxsEmailsList401Response": {
        "title": "get a mailbox's emails list 401 response",
        "required": [
          "success",
          "message"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "No token provided."
          }
        },
        "example": {
          "success": false,
          "message": "No token provided."
        }
      },
...
    }
  },
  "security": [],
...
}